refactor(node): move attestation context into settlement module
