+++ crates/node/src/role.rs
@@ -51,10 +51,9 @@ use crate::{
     EngineExit, ProductionPermit, ZoneEngine, ZoneSequencerAddOnsConfig,
     follower::{BlockSyncP2p, FollowerBlockSync, FollowerBlockSyncContext, PeerTipRegistry},
     replication::{
-        AttestationContext, BroadcasterShutdown, broadcast_persisted_blocks,
-        collect_follower_settlement_signatures,
+        BroadcasterShutdown, broadcast_persisted_blocks, collect_follower_settlement_signatures,
     },
-    settlement_attestation::collect_leader_settlements,
+    settlement_attestation::{AttestationContext, collect_leader_settlements},
     tx_forwarding::{forward_new_transactions, insert_forwarded_transactions},
 };
